requirements.txt
APScheduler==3.6.3
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
clearbit==0.1.7
Click==7.0
click-plugins==1.1.1
colorama==0.4.3
cryptography==3.4.4
geographiclib==1.50
geopy==2.1.0
gitdb==4.0.5
GitPython==3.1.12
html5lib==1.1
idna==2.8
instagram-private-api==1.6.0.0
isort==5.7.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
phonenumbers==8.12.18
Pillow==8.1.0
prettytable==0.7.2
pycparser==2.20
pylint==2.6.0
PySocks==1.7.1
python-telegram-bot==13.2
pytz==2021.1
requests==2.22.0
requests-toolbelt==0.9.1
selenium==3.141.0
shodan==1.21.2
six==1.15.0
smmap==3.0.5
socks==0
soupsieve==2.2
style==1.1.0
tbselenium==0.5.3
toml==0.10.2
tornado==6.1
tzlocal==2.1
update==0.0.1
urllib3==1.25.7
watchdog==2.0.0
webencodings==0.5.1
wrapt==1.12.1
XlsxWriter==1.2.7
